spl: Bump to newest SPL crates for v2.1 release (#3431)
* spl: Bump to newest SPL crates for v2.1 release

#### Problem

A lot of builds are currently broken due to the incompatibility between
v2.0 and v2.1. All of the SPL crates have been bumped to use the Solana
v2.1 crates, but it isn't being used in the monorepo.

#### Summary of changes

Bump the SPL crates to their newest versions, and correctly parse the
new token-2022 instructions and account types.

The trickiest changes are the confidential transfer / mint / burn
instruction parsing, otherwise the changes are straightforward.

* Address feedback

* Improve parsing logic per discussion

* Parse initialize instructions correctly

* Add comments for parsing

* Improve account labels

* Update lockfile from rebase

+++ transaction-status/src/parse_token/extension/token_group.rs
@@ -19,7 +19,7 @@ pub(in crate::parse_token) fn parse_token_group_instruction(
             } = group;
             let value = json!({
                 "group": account_keys[account_indexes[0] as usize].to_string(),
-                "maxSize": u32::from(*max_size),
+                "maxSize": u64::from(*max_size),
                 "mint": account_keys[account_indexes[1] as usize].to_string(),
                 "mintAuthority": account_keys[account_indexes[2] as usize].to_string(),
                 "updateAuthority": Option::<Pubkey>::from(*update_authority).map(|v| v.to_string())
@@ -34,7 +34,7 @@ pub(in crate::parse_token) fn parse_token_group_instruction(
             let UpdateGroupMaxSize { max_size } = update;
             let value = json!({
                 "group": account_keys[account_indexes[0] as usize].to_string(),
-                "maxSize": u32::from(*max_size),
+                "maxSize": u64::from(*max_size),
                 "updateAuthority": account_keys[account_indexes[1] as usize].to_string(),
             });
             Ok(ParsedInstructionEnum {
